units.py: real and complex upcasting, take 2

+++ tamasis/tests/test_unit.py
@@ -143,7 +143,7 @@ b[0]=1
 if a[0] != b[0]: raise TestFailure()
 
 
-a = Tod([10,10], nsamples=(1,1), unit='m')
+a = Tod([10.,10.], nsamples=(1,1), unit='m')
 b = Quantity(a)
 if a.unit != b.unit: raise TestFailure()
 
@@ -180,13 +180,13 @@ if b.unit != 'Jy^2': raise TestFailure()
 
 # test upcasting
 if Quantity(1).dtype.type is not numpy.float64: raise TestFailure()
-if Quantity(numpy.float32(1.)).dtype.type is not numpy.float64: raise TestFailure()
+if Quantity(numpy.float32(1.)).dtype.type is not numpy.float32: raise TestFailure()
 if Quantity(1.).dtype.type is not numpy.float64: raise TestFailure()
 if Quantity(complex(1,0)).dtype.type is not numpy.complex128: raise TestFailure()
-if Quantity(numpy.complex64(1.)).dtype.type is not numpy.complex128: raise TestFailure()
+if Quantity(numpy.complex64(1.)).dtype.type is not numpy.complex64: raise TestFailure()
 if Quantity(numpy.complex128(1.)).dtype.type is not numpy.complex128: raise TestFailure()
 if Quantity(numpy.array(complex(1,0))).dtype.type is not numpy.complex128: raise TestFailure()
-if Quantity(numpy.array(numpy.complex64(1.))).dtype.type is not numpy.complex128: raise TestFailure()
+if Quantity(numpy.array(numpy.complex64(1.))).dtype.type is not numpy.complex64: raise TestFailure()
 if Quantity(numpy.array(numpy.complex128(1.))).dtype.type is not numpy.complex128: raise TestFailure()
 
 print 'OK.'
